Progress bars hacky fixes (#200)
* Progress bars hacky fixes

* Remove  blinking on 100%
